requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
scrapy==2.11.1
crochet==2.0.0
python-dotenv==1.0.1
PySocks==1.7.1
requests==2.32.3
playwright==1.47.2
requests==2.32.3
